refactor: Add edit task functionality to improve task management
This commit adds the functionality to edit tasks in the task management system. It introduces a new modal, "Edit Task Modal", which allows users to edit the name of a task. The changes made in the modal are then sent to the server to update the task. This enhancement improves the user experience by providing the ability to modify task names without having to delete and recreate tasks.