feat: Add "Edit Task" functionality to task management system
This commit introduces the "Edit Task" functionality to the task management system. It includes the addition of an "Edit Task Modal" that allows users to modify the name of a task. The changes made in the modal are then sent to the server to update the task. This enhancement improves the user experience by providing the ability to edit task names without the need to delete and recreate tasks.