chore: Update home page to display kid's savings amount
This commit updates the home page to display the kid's savings amount. It adds a new element with the ID "kid-savings" and sets its display to "none" by default. It also adds a new element with the ID "savings-amount" to show the actual savings amount. The JavaScript code in home.js sets the display of "kid-savings" to "block" and updates the text content of "savings-amount" with the value from the kid's data. This change allows the kid's savings amount to be shown on the home page.