chore: Update home page to display messages
This commit modifies the home page to display messages. It adds a new section with the ID "messages" and a list with the ID "message-list". The JavaScript code in home.js fetches messages from the server and dynamically populates the list with the received messages. If there are no messages available, a default message is displayed. This change enhances the user experience by providing a way to view and interact with messages on the home page.