chore: Update positioning of joke section on home page
The code changes in this commit update the positioning of the joke section on the home page. The top value of the #joke-section element is modified from 15% to 130px, resulting in a more precise placement of the section on the page. This adjustment improves the overall layout and visual appearance of the home page.